Refactoring - Initiated npm - Enabled ESM module loader - Moved codes to different modules as:   - Logger function to logger file   - Sample data to mock_db   - All classes to their respective places.
