+++ node.js
@@ -1,0 +1,8 @@
+export default class Node {
+    constructor(id, treeId, questionText, answers){
+        this.id = id;
+        this.treeId = treeId;
+        this.questionText = questionText;
+        this.answers = answers;
+    };
+}
